+++ src/main/kotlin/com/example/exposed/OrderLineTable.kt
@@ -3,7 +3,7 @@ package com.example.exposed
 import org.jetbrains.exposed.sql.Table
 
 object OrderLineTable : Table("order_line") {
-    val orderId = long("order_id")
+    val orderId = integer("order_id")
         .uniqueIndex()
         .references(OrderTable.id)
     val price = double("price")
